+++ src/ntshengn_window_module.h
@@ -13,7 +13,7 @@ namespace NtshEngn {
 		WindowModule() : WindowModuleInterface("NutshellEngine GLFW Window Module") {}
 
 		void init();
-		void update(double dt);
+		void update(float dt);
 		void destroy();
 
 		// Opens a new window and returns a unique identifier
